Move function into different file
